implement abstract classes for execute and comparison
